Move compaction methods off *levelManager onto *compactor (stage B)
24 compaction methods (15 executor + 3 picker + 2 pacer accessors +
3 metrics + adjustThrottle) are now methods on *compactor instead of
methods on *levelManager. Receiver changes through:
  lm.X (compactor state)  -> c.X
  lm.X (other lm fields)  -> c.lm.X
  lm.METHOD() (moved set) -> c.METHOD()

Scheduler in compaction.go now holds compactor *compactor instead of
owner *levelManager and calls its methods through c.compactor.X.

10 free helpers (resolveTablesLocked, tableMetaSnapshot, findTableByID,
getKeyRange, tablesToString, buildChangeSet, newDeleteChange,
newCreateChange, iteratorsReversed, tablesStrictlyOrdered) move along
with their callers. Throttle helpers normalizedThrottlePressure and
throttleRateForPressure also move (they support adjustThrottle).

CompactionPacerStats deleted entirely (no callers anywhere).

level_manager.go: 1659 -> 480 LoC. Now contains only true
levelManager concerns: lifecycle, manifest, build, Get, iterators,
flush install, rebuildRangeTombstones, level metrics, max version,
cache metrics, table.Runtime impl.
